+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 - Add support of new syntax for tip command #57
 - Add donate command
 - Check wallet pass phrase on startup
+- Fix bug on update status of tip in history
+- Remove extra decimal on date in history
 
 # v1.0 - 10/06/2017
 - First release